pyproject.toml
[project]
name = "cerberus"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "google-cloud-compute>=1.18.0",
    "google-cloud-billing>=1.13.0",
    "google-cloud-asset>=3.24.0",
    "google-cloud-monitoring>=2.22.0",
    "google-cloud-iam>=2.15.0",
    "google-cloud-resource-manager>=1.12.0",
    "google-cloud-logging>=3.10.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "chromadb>=0.5.0",
    "tenacity>=9.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
filterwarnings = [
    "ignore::FutureWarning:google.api_core",
    "ignore::DeprecationWarning:chromadb",
]
